allow match() to be used for any mock request
