+++ addon/mocks/mock-request.js
@@ -1,4 +1,5 @@
 import { assert } from '@ember/debug';
+import { typeOf } from '@ember/utils';
 import { assign } from '@ember/polyfills';
 import {
   isEmptyObject,
@@ -9,6 +10,47 @@ import {
 } from '../utils/helper-functions';
 import FactoryGuy from '../factory-guy';
 import RequestManager from './request-manager';
+
+/**
+ This is tricky, but the main idea here is:
+
+ #1 Take the keys they want to match and transform them to what the serialized
+ version would be ie. company => company_id
+
+ #2 Take the matchArgs and turn them into a FactoryGuy payload class by
+ FactoryGuy.build(ing) them into a payload
+
+ #3 Wrap the request data into a FactoryGuy payload class
+
+ #4 Go though the keys from #1 and check that both the payloads from #2/#3 have the
+ same values
+
+ @param requestData
+ @returns {boolean} true is no attributes to match or they all match
+ */
+const attributesMatch = function (requestData, modelName, matchParams) {
+  if (isEmptyObject(matchParams)) {
+    return true;
+  }
+
+  let builder = FactoryGuy.fixtureBuilder(modelName);
+
+  // transform they match keys
+  let matchCheckKeys = Object.keys(matchParams).map((key) => {
+    return builder.transformKey(modelName, key);
+  });
+  // build the match args into a JSONPayload class
+  let buildOpts = { serializeMode: true, transformKeys: true };
+  let expectedData = builder.convertForBuild(modelName, matchParams, buildOpts);
+
+  // wrap request data in a JSONPayload class
+  builder.wrapPayload(modelName, requestData);
+
+  // success if all values match
+  return matchCheckKeys.every((key) => {
+    return isEquivalent(expectedData.get(key), requestData.get(key));
+  });
+};
 
 export default class {
   constructor() {
@@ -157,6 +199,17 @@ export default class {
   }
 
   extraRequestMatches(request) {
+    if (this.matchArgs) {
+      let requestBody = request.requestBody;
+      if (typeOf(requestBody) === 'string') {
+        requestBody = JSON.parse(requestBody);
+      }
+      if (typeOf(this.matchArgs) === 'function') {
+        return this.matchArgs(requestBody);
+      } else {
+        return attributesMatch(requestBody, this.modelName, this.matchArgs);
+      }
+    }
     return this.paramsMatch(request);
   }
 
@@ -177,6 +230,11 @@ export default class {
     }
 
     return true;
+  }
+
+  match(matches) {
+    this.matchArgs = matches;
+    return this;
   }
 
   // mockId holds the url for this mock request
