+++ addon/mocks/mock-update-request.js
@@ -2,18 +2,16 @@ import { assert } from '@ember/debug';
 import { assign } from '@ember/polyfills';
 import FactoryGuy from '../factory-guy';
 import MockStoreRequest from './mock-store-request';
-import AttributeMatcher from './attribute-matcher';
 import MaybeIdUrlMatch from './maybe-id-url-match';
 
 export default class MockUpdateRequest extends MaybeIdUrlMatch(
-  AttributeMatcher(MockStoreRequest)
+  MockStoreRequest
 ) {
   constructor(modelName, { id, model } = {}) {
     super(modelName, 'updateRecord');
     this.id = id;
     this.model = model;
     this.returnArgs = {};
-    this.matchArgs = {};
     this.setupHandler();
   }
 
@@ -44,6 +42,28 @@ export default class MockUpdateRequest extends MaybeIdUrlMatch(
     this.returnArgs = returns.attrs;
     this.add = returns.add;
     return this;
+  }
+
+  /**
+   Update and Create mocks can accept 2 return keys 'attrs' and 'add'
+
+   @param options
+   @returns {Array}
+   */
+  validateReturnsOptions(options) {
+    const responseKeys = Object.keys(options),
+      validReturnsKeys = ['attrs', 'add'],
+      invalidKeys = responseKeys.filter(
+        (key) => !validReturnsKeys.includes(key)
+      );
+
+    assert(
+      `[ember-data-factory-guy] You passed invalid keys for 'returns' function.
+      Valid keys are ${validReturnsKeys}. You used these invalid keys: ${invalidKeys}`,
+      invalidKeys.length === 0
+    );
+
+    return responseKeys;
   }
 
   /**
